Moving JS event epoch.comment.posted
Event is currently being triggered twice, sending a different comment ID each time. The first comment ID is the temp ID assigned, and the second comment ID is the real comment ID.

Goal is to have the event fire only once when the new comment is successfully inserted into the DOM and the actual comment ID is available.