fix maybe_verify_origin function
The maybe_verify_origin has a comparison that makes no sense. It is comparing a hostname to a full URL which could never match. This patch parses the home_url() and then compares the host value of that to the host value from the parsed url of the request’s referrer to compare apples to apples.